Asio
Add bitcoin params

Add connman, node, banman

Implement connect

1

Serde

+++ include/veriblock/pop/time.hpp
@@ -15,12 +15,13 @@ namespace altintegration {
  * Set mock time for usage in unit tests. 0 disables mock time.
  * @param mocktime
  */
-void setMockTime(uint32_t mocktime);
+void setMockTime(int64_t mocktime);
 
 //! Get current mock time.
 uint32_t getMockTime();
 
-//! Get current time as 4 bytes. If mock time is set, returns mock time.
+//! Get current time as 4 bytes (as seconds). If mock time is set, returns mock
+//! time.
 uint32_t currentTimestamp4();
 
 }  // namespace altintegration
